requirements.txt
discord.py==1.3.4
psutil==5.7.2
python-dotenv==0.14.0